Changed Context-Attribute SortNumber

+++ boephotoarchiveService/DataObjects/Image.cs
@@ -6,7 +6,7 @@ namespace boephotoarchiveService.DataObjects
     {
 
         public string UserID { get; set; }
-        public float Lat { get; set; }
-        public float Lon { get; set; }
+        public double Lat { get; set; }
+        public double Lon { get; set; }
     }
 }
